requirements.txt
python-dotenv
pandas
numpy
pymongo==4.6.0
certifi
scikit-learn
datetime

##-e .